Move Config from commons.core into model.config
We are going to make Config a part of the `Model`.

In preparation for that, move Config into the `model` subpackage.
