+++ src/main/java/seedu/address/commons/util/ConfigUtil.java
@@ -5,9 +5,9 @@ import java.io.IOException;
 import java.util.Optional;
 import java.util.logging.Logger;
 
-import seedu.address.commons.core.Config;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.commons.exceptions.DataConversionException;
+import seedu.address.model.config.Config;
 
 /**
  * A class for accessing the Config File.
