+++ src/test/java/seedu/address/commons/util/ConfigUtilTest.java
@@ -13,8 +13,8 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.junit.rules.TemporaryFolder;
 
-import seedu.address.commons.core.Config;
 import seedu.address.commons.exceptions.DataConversionException;
+import seedu.address.model.config.Config;
 
 public class ConfigUtilTest {
