+++ src/test/java/seedu/address/model/config/ConfigTest.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.core;
+package seedu.address.model.config;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
